send telegram text messages added

+++ app/src/main/java/com/example/tucha/network/model/telegram/TelegramUpdate.kt
@@ -12,8 +12,8 @@ data class TelegramUpdate(
 fun List<TelegramUpdate>.asDialogDatabaseModel(): List<DatabaseDialog> {
     return map {
         DatabaseDialog(
-            id = it.message.user.id,
-            userId = it.message.user.id,
+            id = it.message.user.id.toInt(),
+            userId = it.message.user.id.toInt(),
             unreadCount = null,
             lastMessageText = it.message.text,
             lastMessageDate = it.message.date,
@@ -25,7 +25,7 @@ fun List<TelegramUpdate>.asDialogDatabaseModel(): List<DatabaseDialog> {
 fun List<TelegramUpdate>.asProfileDatabaseModel(): List<DatabaseProfile> {
     return map {
         DatabaseProfile(
-            id = it.message.user.id,
+            id = it.message.user.id.toInt(),
             firstName = it.message.user.firstName
         )
     }
@@ -35,7 +35,7 @@ fun List<TelegramUpdate>.asMessageDatabaseModel(): List<DatabaseMessage> {
     return map {
         DatabaseMessage(
             id = it.message.id,
-            userId = it.message.user.id,
+            userId = it.message.user.id.toInt(),
             out = 0,
             date = it.message.date,
             text = it.message.text
