+++ app/src/main/java/com/example/tucha/network/model/telegram/TelegramUser.kt
@@ -3,6 +3,6 @@ package com.example.tucha.network.model.telegram
 import com.squareup.moshi.Json
 
 data class TelegramUser(
-    @Json(name = "id") val id: Int,
+    @Json(name = "id") val id: Long,
     @Json(name = "first_name") val firstName: String
 )
